pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "google-auth>=2.40.2",
    "gspread>=6.2.1",
    "pdf2image>=1.17.0",
    "requests>=2.32.3",
]
